+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/integrations/emi/CommonEntrypoint.kt
@@ -36,7 +36,7 @@ object CommonEntrypoint {
             )
         }
         if (!stack.itemStack.isEmpty) {
-            val base = LuaRepresentation.forItemStack(stack.itemStack, RepresentationMode.BASE)
+            val base = LuaRepresentation.forItemStack(stack.itemStack, RepresentationMode.DETAILED)
             base["type"] = "item"
             base["chance"] = stack.chance
             return base